Refactoring - move CastaliaConfigSpec into Model.scala, with all the other json-config classes
